previewApiLookup restructuring
